+++ tests/test_version.py
@@ -1,4 +1,4 @@
 def test_version_is_defined():
-    from  airflow_indexima import __version__
+    from airflow_indexima import __version__
 
     assert __version__