+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-# 0.0.0 (2019-11-27)
+# 1.0.0 (2019-11-27)
 
 - initial project structure based on [geronimo-iia/template-python](https://github.com/geronimo-iia/template-python)
+- add Hook implementation
+- add Simple Operator
+- add pyhive, ...
+- configure documentation
+- add a way to customize credentials retreival